qualyx-backend/requirements.txt
	
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0          # async PostgreSQL driver
pandas==2.2.3            # CSV processing
python-multipart==0.0.12 # file uploads
httpx==0.27.2            # async HTTP client
google-generativeai==0.8.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-dotenv==1.0.1
aiosmtplib==3.0.1        # async email
